Collect IAM functionality together in code and tests
