+++ src/JustSaying/AwsTools/MessageHandling/IamPolicyExtensions.cs
@@ -3,7 +3,7 @@ using Amazon;
 
 namespace JustSaying.AwsTools.MessageHandling;
 
-internal static class SnsPolicyDetailsIamExtensions
+internal static class IamPolicyExtensions
 {
     internal static string BuildIamPolicyJson(this SnsPolicyDetails policyDetails)
     {
@@ -48,5 +48,44 @@ internal static class SnsPolicyDetailsIamExtensions
         }}
     ]
 }}";
+    }
+
+    internal static string BuildIamPolicyJson(this SqsPolicyDetails policyDetails)
+    {
+        var sid = Guid.NewGuid().ToString().Replace("-", "");
+
+        var resource = policyDetails.QueueArn;
+
+        var topicArnWildcard = string.IsNullOrWhiteSpace(policyDetails.SourceArn)
+            ? "*"
+            : CreateTopicArnWildcard(policyDetails.SourceArn);
+
+        var policyJson = $@"{{
+    ""Version"" : ""2012-10-17"",
+    ""Statement"" : [
+        {{
+            ""Sid"" : ""{sid}"",
+            ""Effect"" : ""Allow"",
+            ""Principal"" : {{
+                ""AWS"" : ""*""
+            }},
+            ""Action""    : ""sqs:SendMessage"",
+            ""Resource""  : ""{resource}"",
+            ""Condition"" : {{
+                ""ArnLike"" : {{
+                    ""aws:SourceArn"" : ""{topicArnWildcard}""
+                }}
+            }}
+        }}
+    ]
+}}";
+        return policyJson;
+    }
+
+    private static string CreateTopicArnWildcard(string topicArn)
+    {
+        var arn = Arn.Parse(topicArn);
+        arn.Resource = "*";
+        return arn.ToString();
     }
 }
