+++ tests/JustSaying.UnitTests/AwsTools/MessageHandling/Iam/Approvals/IamPolicyExtensionsSnsTests.ShouldGenerateApprovedIamPolicy.approved.txt
