+++ tests/JustSaying.UnitTests/AwsTools/MessageHandling/Iam/Approvals/IamPolicyExtensionsSqsTests.ShouldGenerateApprovedIamPolicy.approved.txt
