+++ tests/JustSaying.UnitTests/AwsTools/MessageHandling/Iam/Approvals/IamPolicyExtensionsSqsTests.ShouldGenerateApprovedIamPolicyWithWildcardFromEmptySourceArn.approved.txt
